include: consolidate typerep_pre.h and mpir_typerep.h
There are some build issue due to typerep_pre.h lives in
src/mpi/datatype/typerep/src but it is used outside in mpir_misc.h.

Consolidate the declarations into mpir_typerep.h. Move the declarations
of MPIR_Localcopy family into mpir_typerep.h as well due to its
dependency on struct MPIR_Typerep_req. They don't fit in mpir_misc.h
anyway.

Declare MPIDI_POSIX_rma_req_t in posix_types.h and use void * in
posix_pre.h. This relieves the struggle of header dependency.
