+++ src/include/mpir_misc.h
@@ -52,69 +52,6 @@ typedef enum MPIR_Lang_t {
 
 extern MPL_initlock_t MPIR_init_lock;
 
-#include "typerep_pre.h"        /* needed for MPIR_Typerep_req */
-
-/* FIXME: bad names. Not gpu-specific, confusing with MPIR_Request.
- *        It's a general async handle.
- */
-typedef enum {
-    MPIR_NULL_REQUEST = 0,
-    MPIR_TYPEREP_REQUEST,
-    MPIR_GPU_REQUEST,
-} MPIR_request_type_t;
-
-typedef struct {
-    union {
-        MPIR_Typerep_req y_req;
-        MPL_gpu_request gpu_req;
-    } u;
-    MPIR_request_type_t type;
-} MPIR_gpu_req;
-
-MPL_STATIC_INLINE_PREFIX void MPIR_async_test(MPIR_gpu_req * areq, int *is_done)
-{
-    int err;
-    switch (areq->type) {
-        case MPIR_NULL_REQUEST:
-            /* a dummy, immediately complete */
-            *is_done = 1;
-            break;
-        case MPIR_TYPEREP_REQUEST:
-            MPIR_Typerep_test(areq->u.y_req, is_done);
-            if (*is_done) {
-                areq->type = MPIR_NULL_REQUEST;
-            }
-            break;
-        case MPIR_GPU_REQUEST:
-            err = MPL_gpu_test(&areq->u.gpu_req, is_done);
-            MPIR_Assertp(err == MPL_SUCCESS);
-            if (*is_done) {
-                areq->type = MPIR_NULL_REQUEST;
-            }
-            break;
-        default:
-            MPIR_Assert(0);
-    }
-}
-
-int MPIR_Localcopy(const void *sendbuf, MPI_Aint sendcount, MPI_Datatype sendtype,
-                   void *recvbuf, MPI_Aint recvcount, MPI_Datatype recvtype);
-int MPIR_Ilocalcopy(const void *sendbuf, MPI_Aint sendcount, MPI_Datatype sendtype,
-                    void *recvbuf, MPI_Aint recvcount, MPI_Datatype recvtype,
-                    MPIR_Typerep_req * typerep_req);
-int MPIR_Localcopy_stream(const void *sendbuf, MPI_Aint sendcount, MPI_Datatype sendtype,
-                          void *recvbuf, MPI_Aint recvcount, MPI_Datatype recvtype, void *stream);
-int MPIR_Localcopy_gpu(const void *sendbuf, MPI_Aint sendcount, MPI_Datatype sendtype,
-                       MPI_Aint sendoffset, MPL_pointer_attr_t * sendattr, void *recvbuf,
-                       MPI_Aint recvcount, MPI_Datatype recvtype, MPI_Aint recvoffset,
-                       MPL_pointer_attr_t * recvattr,
-                       MPL_gpu_engine_type_t enginetype, bool commit);
-int MPIR_Ilocalcopy_gpu(const void *sendbuf, MPI_Aint sendcount, MPI_Datatype sendtype,
-                        MPI_Aint sendoffset, MPL_pointer_attr_t * sendattr, void *recvbuf,
-                        MPI_Aint recvcount, MPI_Datatype recvtype, MPI_Aint recvoffset,
-                        MPL_pointer_attr_t * recvattr,
-                        MPL_gpu_engine_type_t enginetype, bool commit, MPIR_gpu_req * req);
-
 /* Contiguous datatype calculates buffer address with `(char *) buf + dt_true_lb`.
  * However, dt_true_lb is treated as ptrdiff_t (signed), and when buf is MPI_BOTTOM
  * and on 32-bit systems, ubsan will warn the latter overflow. Cast to uintptr_t
